Celebrity problem
If A knows B, then A can’t be a celebrity. Discard A, and B may be celebrity.
If A doesn’t know B, then B can’t be a celebrity. Discard B, and A may be celebrity.
Repeat the above two steps till there is only one person.
Ensure the remained person is a celebrity.