level order travesal
      1
     / \
    2   3
   / \   \
  4   5   6
Level order traversal is a way to traverse or visit all the nodes of a tree in a breadth-first manner. It starts at the root of the tree and visits all the nodes at each level before moving on to the next level.

To perform level order traversal, we can use a queue data structure. We first enqueue the root node and then dequeue it and enqueue its children. We continue this process until all nodes have been visited.